dela_manager/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
psycopg2-binary
pydantic
python-jose
bcrypt
httpx
pytest
